Make `get_block_contrast()` return an `int`
So far the function returned a `numbers.Number`, depending on the `dtype` of
the block being passed.

This commit makes the function explicitely return an `int`.

The contrast of the block is the difference between intensities, which
semantically are integers. The function returning floats in our case was a side
effect of converting our image data to double precision. With this change, the
return value is more precisely defined and semantically sensible. In addition,
it may also be used as an index into the `width_jnb` array ;)