Stitch together a `sobel` implementation
Try to emulate MATLAB's results. The current approach does not work well for
all images:

```
>>> zip(reference_sharpnesses, my_sharpnesses)
[(0.1349, 0.25724217844727693),
 (0.3418, 0.32393679282091303),
 (0.0161, 0.10913642052565707),
 (0.7479, 0.75480910641802468),
 (0.00049652, 0.36764174792343807),
 (0.0677, 0.11278667923342758),
 (0.4452, 0.42657869934024506),
 (0.4842, 0.46370112797896701),
 (0.5726, 0.57193378480060197)]
```

I need to take a look at the images that the operation performs particulary
poorly on. Maybe a property sticks out.

One might still try to cheat and use a `Scharr` operator or do thresholding
with the `median`. Or try another Sobel implementation as a base.