Use only vertical edges
In particularly bad cases (e.g. `img5.bmp`) the sobel detector picked up
too many edges, due to the horizontal components used to calculate the
magnitude.

Therefore this commit considers only vertical edges. As a result, the
calculation of the magnitude was reduced to squaring the horizontal
derivatives (for the vertical derivatives were dropped, and continuing
to take the square root would just have cancelled that out).